split test_validate_schema.py
